contest ques added

+++ Contest/Arrays/ReadingPages.cpp
@@ -1,15 +1,15 @@
 #include<iostream>
 using namespace std;
 int main(){
-	long long int n,m;
+	int n,m;
 	cin>>n>>m;
-	long long int p[1000];
-	for(long long int i=0;i<n;i++){
+	int p[1000];
+	for(int i=0;i<n;i++){
 		cin>>p[i];
 	}
-	long long int dec=m-1;
-	long long int max=0;
-	for(long long int i=0;i<n;i++){
+	int dec=m-1;
+	int max=0;
+	for(int i=0;i<n;i++){
 		max+=p[i];
 	}
 	while(p[0]>0){
